move OffsetQuantity into its own coulomb-offset package
