+++ build.sbt
@@ -36,9 +36,16 @@ lazy val coulomb_si_units = (project in file("coulomb-si-units"))
   .settings(commonSettings :_*)
   .settings(docDepSettings :_*)
 
+lazy val coulomb_offset = (project in file ("coulomb-offset"))
+  .aggregate(coulomb)
+  .dependsOn(coulomb)
+  .settings(name := "coulomb-offset")
+  .settings(commonSettings :_*)
+  .settings(docDepSettings :_*)
+
 lazy val coulomb_temp_units = (project in file("coulomb-temp-units"))
-  .aggregate(coulomb, coulomb_si_units)
-  .dependsOn(coulomb, coulomb_si_units)
+  .aggregate(coulomb, coulomb_offset, coulomb_si_units)
+  .dependsOn(coulomb, coulomb_offset, coulomb_si_units)
   .settings(name := "coulomb-temp-units")
   .settings(commonSettings :_*)
   .settings(docDepSettings :_*)
@@ -127,8 +134,8 @@ lazy val coulomb_pureconfig = (project in file("coulomb-pureconfig"))
   .settings(libraryDependencies ++= coulombPureConfigDeps)
 
 lazy val coulomb_tests = (project in file("coulomb-tests"))
-  .aggregate(coulomb, coulomb_si_units, coulomb_mks_units, coulomb_accepted_units, coulomb_time_units, coulomb_info_units, coulomb_customary_units, coulomb_temp_units, coulomb_parser, coulomb_typesafe_config, coulomb_avro, coulomb_pureconfig)
-  .dependsOn(coulomb, coulomb_si_units, coulomb_mks_units, coulomb_accepted_units, coulomb_time_units, coulomb_info_units, coulomb_customary_units, coulomb_temp_units, coulomb_parser, coulomb_typesafe_config, coulomb_avro, coulomb_pureconfig)
+  .aggregate(coulomb, coulomb_offset, coulomb_si_units, coulomb_mks_units, coulomb_accepted_units, coulomb_time_units, coulomb_info_units, coulomb_customary_units, coulomb_temp_units, coulomb_parser, coulomb_typesafe_config, coulomb_avro, coulomb_pureconfig)
+  .dependsOn(coulomb, coulomb_offset, coulomb_si_units, coulomb_mks_units, coulomb_accepted_units, coulomb_time_units, coulomb_info_units, coulomb_customary_units, coulomb_temp_units, coulomb_parser, coulomb_typesafe_config, coulomb_avro, coulomb_pureconfig)
   .settings(name := "coulomb-tests")
   .settings(commonSettings :_*)
   .settings(libraryDependencies ++= coulombAvroDeps)
@@ -137,8 +144,8 @@ lazy val coulomb_tests = (project in file("coulomb-tests"))
   .settings(libraryDependencies ++= coulombPureConfigDeps)
 
 lazy val coulomb_docs = (project in file("."))
-  .aggregate(coulomb, coulomb_si_units, coulomb_mks_units, coulomb_accepted_units, coulomb_time_units, coulomb_info_units, coulomb_customary_units, coulomb_temp_units, coulomb_parser, coulomb_typesafe_config, coulomb_avro, coulomb_pureconfig)
-  .dependsOn(coulomb, coulomb_si_units, coulomb_mks_units, coulomb_accepted_units, coulomb_time_units, coulomb_info_units, coulomb_customary_units, coulomb_temp_units, coulomb_parser, coulomb_typesafe_config, coulomb_avro, coulomb_pureconfig)
+  .aggregate(coulomb, coulomb_offset, coulomb_si_units, coulomb_mks_units, coulomb_accepted_units, coulomb_time_units, coulomb_info_units, coulomb_customary_units, coulomb_temp_units, coulomb_parser, coulomb_typesafe_config, coulomb_avro, coulomb_pureconfig)
+  .dependsOn(coulomb, coulomb_offset, coulomb_si_units, coulomb_mks_units, coulomb_accepted_units, coulomb_time_units, coulomb_info_units, coulomb_customary_units, coulomb_temp_units, coulomb_parser, coulomb_typesafe_config, coulomb_avro, coulomb_pureconfig)
   .settings(name := "coulomb-docs")
   .settings(commonSettings :_*)
