+++ coulomb-offset/src/main/scala/coulomb/offset/OffsetQuantity.scala
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package coulomb.temp
+package coulomb.offset
 
 import scala.language.implicitConversions
